+++ inst/shiny/global.R
@@ -23,3 +23,7 @@ pathData <- setNames(
   list.dirs(pathDBs, full.names = TRUE, recursive = FALSE), 
   list.dirs(pathDBs, full.names = FALSE, recursive = FALSE)
 )
+
+# The database could be loaded here to save memory
+# Instead, we want to let the user choose the database
+# Then it has to be loaded in server.R, once for each user.